service_1/requirements.txt
aws-cdk.core

aws-cdk.aws_dynamodb

boto3
botocore
